Move most fractal UI and logic to instances
