+++ src/gui/ui/mod.rs
@@ -1,20 +1,25 @@
+mod instance;
 mod viewer;
 
 use crate::{
     generator::view::View,
-    gui::{keyboard::KeyboardTracker, ui::viewer::FractalViewer},
-    util::result::ResultExt,
+    gui::{
+        keyboard::KeyboardTracker,
+        ui::instance::{UIInstance, UIInstanceCreationContext},
+    },
 };
-use egui::{vec2, CtxRef, DragValue, ProgressBar};
+use egui::{CtxRef, Layout};
 use egui_wgpu_backend::RenderPass;
-use std::borrow::Cow;
-use wgpu::Device;
+use std::sync::Arc;
+use wgpu::{Device, Queue};
 use winit::event::VirtualKeyCode;
 
-const DEFAULT_GENERATION_MESSAGE: &str = "Not Generating";
-
 /// Struct specifically devoted to UI rendering and state.
-pub struct UIState {
+pub struct FractalRSUI {
+    // needed for creating new instances
+    _device: Arc<Device>,
+    _queue: Arc<Queue>,
+
     // application flow controls
     pub close_requested: bool,
 
@@ -23,30 +28,19 @@ pub struct UIState {
     pub request_fullscreen: bool,
 
     // open windows
-    pub show_generator_controls: bool,
-    pub show_viewer_controls: bool,
-    pub show_ui_settings: bool,
+    show_ui_settings: bool,
 
-    // generator controls
-    pub generate_fractal: bool,
-    pub generation_fraction: f32,
-    pub generation_message: Cow<'static, str>,
-    pub edit_fractal_width: usize,
-    pub edit_fractal_height: usize,
-    pub edit_fractal_plane_width: f32,
-    pub edit_fractal_plane_centered: bool,
-    pub edit_fractal_plane_center_x: f32,
-    pub edit_fractal_plane_center_y: f32,
-    pub fractal_view: View,
-
-    // fractal viewers
-    pub julia_viewer: FractalViewer,
+    // instances
+    instances: Vec<UIInstance>,
+    current_instance: usize,
 }
 
 /// Struct containing context passed when creating UIState.
 pub struct UICreationContext<'a> {
     /// Device reference.
-    pub device: &'a Device,
+    pub device: Arc<Device>,
+    /// Queue reference.
+    pub queue: Arc<Queue>,
     /// WGPU Egui Render Pass reference for managing textures.
     pub render_pass: &'a mut RenderPass,
     /// Fractal view settings at the time of UI state creation.
@@ -55,57 +49,56 @@ pub struct UICreationContext<'a> {
 
 /// Struct containing context passed to the UI render function.
 pub struct UIRenderContext<'a> {
-    /// Device reference.
-    pub device: &'a Device,
     /// Egui context reference.
     pub ctx: &'a CtxRef,
     /// WGPU Egui Render Pass reference for managing textures.
     pub render_pass: &'a mut RenderPass,
-    /// Whether the fractal generator instance is currently running.
-    pub not_running: bool,
     /// Tracker for currently pressed keys.
     pub keys: &'a KeyboardTracker,
 }
 
-impl UIState {
+impl FractalRSUI {
     /// Create a new UIState, making sure to initialize all required textures
     /// and such.
-    pub fn new(ctx: &mut UICreationContext) -> UIState {
-        let plane_width =
-            ctx.initial_fractal_view.image_width as f32 * ctx.initial_fractal_view.image_scale_x;
-        let plane_height =
-            ctx.initial_fractal_view.image_height as f32 * ctx.initial_fractal_view.image_scale_y;
-        let center_x = ctx.initial_fractal_view.plane_start_x + plane_width / 2.0;
-        let center_y = ctx.initial_fractal_view.plane_start_y + plane_height / 2.0;
-
-        UIState {
+    pub async fn new(ctx: UICreationContext<'_>) -> FractalRSUI {
+        let first_instance = UIInstance::new(UIInstanceCreationContext {
+            name: "Fractal 1",
+            device: ctx.device.clone(),
+            queue: ctx.queue.clone(),
+            render_pass: ctx.render_pass,
+            initial_fractal_view: ctx.initial_fractal_view,
+        })
+        .await;
+        FractalRSUI {
+            _device: ctx.device,
+            _queue: ctx.queue,
             close_requested: false,
             previous_fullscreen: false,
             request_fullscreen: false,
-            show_generator_controls: true,
-            show_viewer_controls: true,
             show_ui_settings: false,
-            generate_fractal: false,
-            generation_fraction: 0.0,
-            generation_message: Cow::Borrowed(DEFAULT_GENERATION_MESSAGE),
-            edit_fractal_width: ctx.initial_fractal_view.image_width,
-            edit_fractal_height: ctx.initial_fractal_view.image_height,
-            edit_fractal_plane_width: plane_width,
-            edit_fractal_plane_centered: center_x == 0.0 && center_y == 0.0,
-            edit_fractal_plane_center_x: center_x,
-            edit_fractal_plane_center_y: center_y,
-            fractal_view: ctx.initial_fractal_view,
-            julia_viewer: FractalViewer::new(ctx.device, ctx.render_pass, ctx.initial_fractal_view),
+            instances: vec![first_instance],
+            current_instance: 0,
+        }
+    }
+
+    /// Update things associated with the UI but that do not involve rendering.
+    pub fn update(&mut self) {
+        // Update all the instances, even the ones that are not currently being
+        // rendered.
+        for instance in self.instances.iter_mut() {
+            instance.update();
         }
     }
 
     /// Render the current UI state to the Egui context.
     pub fn draw(&mut self, ctx: &mut UIRenderContext) {
         self.handle_keyboard_shortcuts(ctx);
-        self.draw_menubar(ctx);
-        self.draw_fractal_viewers(ctx);
-        self.draw_generator_controls(ctx);
-        self.draw_viewer_controls(ctx);
+        self.draw_top_bar(ctx);
+        if let Some(instance) = self.open_instance() {
+            instance.draw(ctx);
+        } else {
+            self.draw_empty_content(ctx);
+        }
         self.draw_misc_windows(ctx);
     }
 
@@ -123,7 +116,7 @@ impl UIState {
         }
     }
 
-    fn draw_menubar(&mut self, ctx: &UIRenderContext) {
+    fn draw_top_bar(&mut self, ctx: &UIRenderContext) {
         egui::TopBottomPanel::top("Menu Bar").show(ctx.ctx, |ui| {
             egui::menu::bar(ui, |ui| {
                 egui::menu::menu(ui, "File", |ui| {
@@ -134,146 +127,44 @@ impl UIState {
                 egui::menu::menu(ui, "Window", |ui| {
                     ui.checkbox(&mut self.request_fullscreen, "Fullscreen");
                     ui.separator();
-                    ui.checkbox(&mut self.show_generator_controls, "Generator Controls");
-                    ui.checkbox(&mut self.show_viewer_controls, "Viewer Controls");
+                    if let Some(instance) = self.open_instance() {
+                        instance.draw_window_options(ctx, ui);
+                        ui.separator();
+                    }
                     ui.checkbox(&mut self.show_ui_settings, "UI Settings");
                 });
             });
+            ui.separator();
+            ui.with_layout(Layout::right_to_left(), |ui| {
+                ui.add_enabled_ui(!self.instances.is_empty(), |ui| {
+                    if ui.button("X").clicked() {
+                        if self.current_instance < self.instances.len() {
+                            if self.current_instance < 0 {
+                                self.current_instance = 0;
+                            }
+                            self.instances.remove(self.current_instance);
+                            if self.current_instance > 0 {
+                                self.current_instance -= 1;
+                            }
+                        } else {
+                            self.current_instance = 0;
+                        }
+                    }
+                });
+                ui.with_layout(Layout::left_to_right(), |ui| {
+                    for (index, instance) in self.instances.iter().enumerate() {
+                        let res = ui.button(&instance.name);
+                        if res.clicked() {
+                            self.current_instance = index;
+                        }
+                    }
+                });
+            });
         });
     }
 
-    fn draw_fractal_viewers(&mut self, ctx: &UIRenderContext) {
-        egui::CentralPanel::default().show(ctx.ctx, |ui| {
-            let available_size = ui.available_size_before_wrap();
-            ui.add_sized(
-                available_size,
-                self.julia_viewer.widget().max_size_override(available_size),
-            );
-        });
-    }
-
-    fn draw_generator_controls(&mut self, ctx: &mut UIRenderContext) {
-        egui::Window::new("Generator Controls")
-            .default_size([250.0, 500.0])
-            .open(&mut self.show_generator_controls)
-            .show(ctx.ctx, |ui| {
-                ui.add_enabled_ui(ctx.not_running, |ui| {
-                    if ui.button("Generate!").clicked() {
-                        self.generate_fractal = true;
-                    }
-                });
-
-                ui.add(ProgressBar::new(self.generation_fraction).text(&self.generation_message));
-
-                ui.separator();
-
-                // actual generator settings
-                egui::Grid::new("generator settings").show(ui, |ui| {
-                    ui.label("Image Width:");
-                    ui.add(
-                        DragValue::new(&mut self.edit_fractal_width)
-                            .speed(1.0)
-                            .clamp_range(64..=4096),
-                    );
-                    ui.end_row();
-
-                    ui.label("Image Height:");
-                    ui.add(
-                        DragValue::new(&mut self.edit_fractal_height)
-                            .speed(1.0)
-                            .clamp_range(64..=4096),
-                    );
-                    ui.end_row();
-
-                    ui.label("Plane Width:");
-                    ui.add(
-                        DragValue::new(&mut self.edit_fractal_plane_width)
-                            .clamp_range(0.0..=10.0)
-                            .speed(0.03125),
-                    );
-                    ui.end_row();
-
-                    ui.checkbox(
-                        &mut self.edit_fractal_plane_centered,
-                        "Centered at (0 + 0i)",
-                    );
-                    ui.end_row();
-
-                    ui.label("Plane Real Center:");
-                    ui.add_enabled(
-                        !self.edit_fractal_plane_centered,
-                        DragValue::new(&mut self.edit_fractal_plane_center_x)
-                            .clamp_range(-10.0..=10.0)
-                            .speed(0.0625),
-                    );
-                    ui.end_row();
-
-                    ui.label("Plane Imaginary Center:");
-                    ui.add_enabled(
-                        !self.edit_fractal_plane_centered,
-                        DragValue::new(&mut self.edit_fractal_plane_center_y)
-                            .clamp_range(-10.0..=10.0)
-                            .speed(0.0625),
-                    );
-                    ui.end_row();
-                });
-            });
-
-        if self.generate_fractal {
-            self.apply_generator_settings(ctx);
-        }
-    }
-
-    fn apply_generator_settings(&mut self, ctx: &mut UIRenderContext) {
-        // apply fractal size
-        self.fractal_view = if self.edit_fractal_plane_centered {
-            View::new_centered_uniform(
-                self.edit_fractal_width,
-                self.edit_fractal_height,
-                self.edit_fractal_plane_width,
-            )
-        } else {
-            View::new_uniform(
-                self.edit_fractal_width,
-                self.edit_fractal_height,
-                self.edit_fractal_plane_width,
-                self.edit_fractal_plane_center_x,
-                self.edit_fractal_plane_center_y,
-            )
-        };
-        self.julia_viewer
-            .set_fractal_view(ctx.device, ctx.render_pass, self.fractal_view)
-            .on_err(|e| error!("Error resizing fractal image: {:?}", e));
-    }
-
-    fn draw_viewer_controls(&mut self, ctx: &UIRenderContext) {
-        egui::Window::new("Viewer Controls")
-            .default_size([250.0, 500.0])
-            .open(&mut self.show_viewer_controls)
-            .show(&ctx.ctx, |ui| {
-                ui.label("Zoom & Center");
-                ui.horizontal(|ui| {
-                    if ui.button("Zoom 1:1").clicked() {
-                        self.julia_viewer.zoom_1_to_1();
-                    }
-                    if ui.button("Zoom Fit").clicked() {
-                        self.julia_viewer.zoom_fit();
-                    }
-                    if ui.button("Zoom Fill").clicked() {
-                        self.julia_viewer.zoom_fill();
-                    }
-                });
-                if ui.button("Center View").clicked() {
-                    self.julia_viewer.fractal_offset = vec2(0.0, 0.0);
-                }
-
-                ui.separator();
-
-                ui.label("Selection");
-                if ui.button("Deselect Position").clicked() {
-                    self.julia_viewer.selection_pos = None;
-                }
-            });
+    fn draw_empty_content(&mut self, ctx: &UIRenderContext) {
+        egui::CentralPanel::default().show(ctx.ctx, |ui| {});
     }
 
     fn draw_misc_windows(&mut self, ctx: &UIRenderContext) {
@@ -284,5 +175,20 @@ impl UIState {
                     ctx.ctx.settings_ui(ui);
                 });
             });
+    }
+
+    fn open_instance(&mut self) -> Option<&mut UIInstance> {
+        if self.instances.is_empty() {
+            None
+        } else {
+            if self.current_instance >= self.instances.len() {
+                self.current_instance = self.instances.len() - 1;
+            }
+            if self.current_instance < 0 {
+                self.current_instance = 0;
+            }
+
+            self.instances.get_mut(self.current_instance)
+        }
     }
 }
